# [sw] Optimize functions in //sw/device/lib/base:memory