--- a/sw/device/lib/base/memory.c
+++ b/sw/device/lib/base/memory.c
@@ -4,6 +4,8 @@
 
 #include "sw/device/lib/base/memory.h"
 
+#include <stdint.h>
+
 #include "sw/device/lib/base/macros.h"
 
 #ifdef OT_PLATFORM_RV32
@@ -12,25 +14,86 @@
 #define OT_PREFIX_IF_NOT_RV32(name) ot_##name
 #endif
 
-// The symbols below are all provided by the host libc. To avoid linker clashes,
-// they are all marked as `OT_WEAK`.
+static size_t compute_num_leading_bytes(const void *left, const void *right,
+                                        size_t len) {
+  if (len < alignof(uint32_t)) {
+    return len;
+  }
+  const size_t left_ahead = misalignment32_of((uintptr_t)left);
+  const size_t right_ahead = misalignment32_of((uintptr_t)right);
+  if (right == NULL || left_ahead == right_ahead) {
+    return (4 - left_ahead) & 0x3;
+  }
+  return len;
+}
 
-OT_WEAK
+/**
+ * Compute the bounds of the word-aligned region for the given buffers.
+ *
+ * It's more efficient for our memory functions to operate on `uint32_t` values
+ * than individual bytes, but we can only read `uint32_t` values from aligned
+ * addresses. This function effectively breaks the given buffers into three
+ * consecutive chunks: the unaligned "head", the aligned "body", and the
+ * unaligned "tail".
+ *
+ * @param[in] left The memory function's first buffer argument. Cannot be NULL.
+ * @param[in] right The memory function's second buffer argument. May be NULL.
+ * @param[in] len The length in bytes of both `left` and `right.`
+ * @param[out] out_body_offset The start of the body region.
+ * @param[out] out_tail_offset The start of the tail region.
+ */
+static void compute_alignment(const void *left, const void *right, size_t len,
+                              size_t *out_body_offset,
+                              size_t *out_tail_offset) {
+  const size_t num_leading_bytes = compute_num_leading_bytes(left, right, len);
+  *out_body_offset = num_leading_bytes;
+
+  const size_t num_words = (len - num_leading_bytes) / sizeof(uint32_t);
+  *out_tail_offset = num_leading_bytes + num_words * sizeof(uint32_t);
+}
+
+static uint32_t repeat_byte_to_u32(uint8_t byte) {
+  return byte << 24 | byte << 16 | byte << 8 | byte;
+}
+
 void *OT_PREFIX_IF_NOT_RV32(memcpy)(void *restrict dest,
                                     const void *restrict src, size_t len) {
-  uint8_t *dest8 = (uint8_t *)dest;
-  uint8_t *src8 = (uint8_t *)src;
-  for (size_t i = 0; i < len; ++i) {
+  if (dest == NULL || src == NULL) {
+    return dest;
+  }
+  unsigned char *dest8 = (unsigned char *)dest;
+  const unsigned char *src8 = (const unsigned char *)src;
+  size_t body_offset, tail_offset;
+  compute_alignment(dest, src, len, &body_offset, &tail_offset);
+  size_t i = 0;
+  for (; i < body_offset; ++i) {
+    dest8[i] = src8[i];
+  }
+  for (; i < tail_offset; i += sizeof(uint32_t)) {
+    uint32_t word = read_32(&src8[i]);
+    write_32(word, &dest8[i]);
+  }
+  for (; i < len; ++i) {
     dest8[i] = src8[i];
   }
   return dest;
 }
 
-OT_WEAK
 void *OT_PREFIX_IF_NOT_RV32(memset)(void *dest, int value, size_t len) {
-  uint8_t *dest8 = (uint8_t *)dest;
-  uint8_t value8 = (uint8_t)value;
-  for (size_t i = 0; i < len; ++i) {
+  unsigned char *dest8 = (unsigned char *)dest;
+  const uint8_t value8 = (uint8_t)value;
+
+  size_t body_offset, tail_offset;
+  compute_alignment(dest, NULL, len, &body_offset, &tail_offset);
+  size_t i = 0;
+  for (; i < body_offset; ++i) {
+    dest8[i] = value8;
+  }
+  const uint32_t value32 = repeat_byte_to_u32(value8);
+  for (; i < tail_offset; i += sizeof(uint32_t)) {
+    write_32(value32, &dest8[i]);
+  }
+  for (; i < len; ++i) {
     dest8[i] = value8;
   }
   return dest;
@@ -42,12 +105,32 @@ enum {
   kMemCmpGt = 42,
 };
 
-OT_WEAK
 int OT_PREFIX_IF_NOT_RV32(memcmp)(const void *lhs, const void *rhs,
                                   size_t len) {
-  const uint8_t *lhs8 = (uint8_t *)lhs;
-  const uint8_t *rhs8 = (uint8_t *)rhs;
-  for (size_t i = 0; i < len; ++i) {
+  const unsigned char *lhs8 = (const unsigned char *)lhs;
+  const unsigned char *rhs8 = (const unsigned char *)rhs;
+  size_t body_offset, tail_offset;
+  compute_alignment(lhs, rhs, len, &body_offset, &tail_offset);
+  size_t i = 0;
+  for (; i < body_offset; ++i) {
+    if (lhs8[i] < rhs8[i]) {
+      return kMemCmpLt;
+    } else if (lhs8[i] > rhs8[i]) {
+      return kMemCmpGt;
+    }
+  }
+  for (; i < tail_offset; i += sizeof(uint32_t)) {
+    uint32_t word_left = __builtin_bswap32(read_32(&lhs8[i]));
+    uint32_t word_right = __builtin_bswap32(read_32(&rhs8[i]));
+    static_assert(__BYTE_ORDER__ == __ORDER_LITTLE_ENDIAN__,
+                  "memcmp assumes that the system is little endian.");
+    if (word_left < word_right) {
+      return kMemCmpLt;
+    } else if (word_left > word_right) {
+      return kMemCmpGt;
+    }
+  }
+  for (; i < len; ++i) {
     if (lhs8[i] < rhs8[i]) {
       return kMemCmpLt;
     } else if (lhs8[i] > rhs8[i]) {
@@ -57,42 +140,120 @@ int OT_PREFIX_IF_NOT_RV32(memcmp)(const void *lhs, const void *rhs,
   return kMemCmpEq;
 }
 
-OT_WEAK
 int memrcmp(const void *lhs, const void *rhs, size_t len) {
-  const uint8_t *lhs8 = (uint8_t *)lhs;
-  const uint8_t *rhs8 = (uint8_t *)rhs;
-  size_t j;
-  for (size_t i = 0; i < len; ++i) {
-    j = len - 1 - i;
-    if (lhs8[j] < rhs8[j]) {
+  const unsigned char *lhs8 = (const unsigned char *)lhs;
+  const unsigned char *rhs8 = (const unsigned char *)rhs;
+  size_t body_offset, tail_offset;
+  compute_alignment(lhs, rhs, len, &body_offset, &tail_offset);
+  size_t end = len;
+  for (; end > tail_offset; --end) {
+    const size_t i = end - 1;
+    if (lhs8[i] < rhs8[i]) {
       return kMemCmpLt;
-    } else if (lhs8[j] > rhs8[j]) {
+    } else if (lhs8[i] > rhs8[i]) {
+      return kMemCmpGt;
+    }
+  }
+  for (; end > body_offset; end -= sizeof(uint32_t)) {
+    const size_t i = end - sizeof(uint32_t);
+    uint32_t word_left = read_32(&lhs8[i]);
+    uint32_t word_right = read_32(&rhs8[i]);
+    static_assert(__BYTE_ORDER__ == __ORDER_LITTLE_ENDIAN__,
+                  "memrcmp assumes that the system is little endian.");
+    if (word_left < word_right) {
+      return kMemCmpLt;
+    } else if (word_left > word_right) {
+      return kMemCmpGt;
+    }
+  }
+  for (; end > 0; --end) {
+    const size_t i = end - 1;
+    if (lhs8[i] < rhs8[i]) {
+      return kMemCmpLt;
+    } else if (lhs8[i] > rhs8[i]) {
       return kMemCmpGt;
     }
   }
   return kMemCmpEq;
 }
 
-OT_WEAK
 void *OT_PREFIX_IF_NOT_RV32(memchr)(const void *ptr, int value, size_t len) {
-  uint8_t *ptr8 = (uint8_t *)ptr;
-  uint8_t value8 = (uint8_t)value;
-  for (size_t i = 0; i < len; ++i) {
+  const unsigned char *ptr8 = (const unsigned char *)ptr;
+  const uint8_t value8 = (uint8_t)value;
+
+  size_t body_offset, tail_offset;
+  compute_alignment(ptr, NULL, len, &body_offset, &tail_offset);
+  size_t i = 0;
+  for (; i < body_offset; ++i) {
     if (ptr8[i] == value8) {
-      return ptr8 + i;
+      return (void *)&ptr8[i];
+    }
+  }
+  const uint32_t value32 = repeat_byte_to_u32(value8);
+  for (; i < tail_offset; i += sizeof(uint32_t)) {
+    uint32_t word = read_32(&ptr8[i]);
+    uint32_t bits_eq = ~(word ^ value32);
+    static_assert(__BYTE_ORDER__ == __ORDER_LITTLE_ENDIAN__,
+                  "memchr assumes that the system is little endian.");
+    if ((bits_eq & UINT8_MAX) == UINT8_MAX) {
+      return (void *)&ptr8[i];
+    }
+    if (((bits_eq >> 8) & UINT8_MAX) == UINT8_MAX) {
+      return (void *)&ptr8[i + 1];
+    }
+    if (((bits_eq >> 16) & UINT8_MAX) == UINT8_MAX) {
+      return (void *)&ptr8[i + 2];
+    }
+    if (((bits_eq >> 24) & UINT8_MAX) == UINT8_MAX) {
+      return (void *)&ptr8[i + 3];
+    }
+  }
+  for (; i < len; ++i) {
+    if (ptr8[i] == value8) {
+      return (void *)&ptr8[i];
     }
   }
   return NULL;
 }
 
-OT_WEAK
 void *OT_PREFIX_IF_NOT_RV32(memrchr)(const void *ptr, int value, size_t len) {
-  uint8_t *ptr8 = (uint8_t *)ptr;
-  uint8_t value8 = (uint8_t)value;
-  for (size_t i = 0; i < len; ++i) {
-    size_t idx = len - i - 1;
-    if (ptr8[idx] == value8) {
-      return ptr8 + idx;
+  const unsigned char *ptr8 = (const unsigned char *)ptr;
+  const uint8_t value8 = (uint8_t)value;
+
+  size_t body_offset, tail_offset;
+  compute_alignment(ptr, NULL, len, &body_offset, &tail_offset);
+
+  size_t end = len;
+  for (; end > tail_offset; --end) {
+    const size_t i = end - 1;
+    if (ptr8[i] == value8) {
+      return (void *)&ptr8[i];
+    }
+  }
+  const uint32_t value32 = repeat_byte_to_u32(value8);
+  for (; end > body_offset; end -= sizeof(uint32_t)) {
+    const size_t i = end - sizeof(uint32_t);
+    uint32_t word = read_32(&ptr8[i]);
+    uint32_t bits_eq = ~(word ^ value32);
+    static_assert(__BYTE_ORDER__ == __ORDER_LITTLE_ENDIAN__,
+                  "memrchr assumes that the system is little endian.");
+    if (((bits_eq >> 24) & UINT8_MAX) == UINT8_MAX) {
+      return (void *)&ptr8[i + 3];
+    }
+    if (((bits_eq >> 16) & UINT8_MAX) == UINT8_MAX) {
+      return (void *)&ptr8[i + 2];
+    }
+    if (((bits_eq >> 8) & UINT8_MAX) == UINT8_MAX) {
+      return (void *)&ptr8[i + 1];
+    }
+    if ((bits_eq & UINT8_MAX) == UINT8_MAX) {
+      return (void *)&ptr8[i];
+    }
+  }
+  for (; end > 0; --end) {
+    const size_t i = end - 1;
+    if (ptr8[i] == value8) {
+      return (void *)&ptr8[i];
     }
   }
   return NULL;