This commit optimizes the implementations of memcpy() and friends. They
now operate on 32-bit words instead of bytes.

To compare the performance impact of my optimizations, I ran the
perftest without updating the cycle count expectations.

```
./bazelisk.sh test  --test_output=streamed //sw/device/lib/base:memory_perftest_fpga_cw310

  W00001 memory_perftest.c:246] memcpy:
    Expected:            210280 cycles
    Actual:               33270 cycles
    Actual/Expected:         15%

  W00002 memory_perftest.c:246] memcpy_zeroes:
    Expected:            210280 cycles
    Actual:               33270 cycles
    Actual/Expected:         15%

  W00003 memory_perftest.c:246] memset:
    Expected:            130310 cycles
    Actual:               23200 cycles
    Actual/Expected:         17%

  W00004 memory_perftest.c:246] memset_zeroes:
    Expected:            130310 cycles
    Actual:               23200 cycles
    Actual/Expected:         17%

  W00005 memory_perftest.c:246] memcmp_pathological:
    Expected:            210310 cycles
    Actual:              110740 cycles
    Actual/Expected:         52%

  W00006 memory_perftest.c:246] memcmp_zeroes:
    Expected:            210280 cycles
    Actual:              110740 cycles
    Actual/Expected:         52%

  W00007 memory_perftest.c:246] memrcmp_pathological:
    Expected:            200330 cycles
    Actual:               50740 cycles
    Actual/Expected:         25%

  W00008 memory_perftest.c:246] memrcmp_zeroes:
    Expected:            200300 cycles
    Actual:               50850 cycles
    Actual/Expected:         25%

  W00009 memory_perftest.c:246] memchr_pathological:
    Expected:             12820 cycles
    Actual:                7250 cycles
    Actual/Expected:         56%

  W00010 memory_perftest.c:246] memrchr_pathological:
    Expected:             47890 cycles
    Actual:               23850 cycles
    Actual/Expected:         49%
```
Issue https://github.com/lowRISC/opentitan/issues/13471